pyproject.toml
[tool.poetry]
name = "sportslabkit"
version = "0.0.1"
description = "A Python package for sports analytics."
authors = ["Atom Scott"]
license = "GPL-3.0-or-later"

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^8.0.6"

[build-system]
requires = ["wheel", "setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py310']

[tool.ruff]
# Never enforce `E501` (line length violations).
ignore = ["C901", "E501", "E741"]
select = ["C", "E", "F", "I", "W"]
line-length = 119

# Ignore import violations in all `__init__.py` files.
[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.isort]
lines-after-imports = 2
known-first-party = ["sportslabkit"]

[tool.mypy]
ignore_missing_imports = true
follow_imports = "skip"
strict_optional = true

[tool.pep257]
ignore = ["D101", "D105", "D200", "D202", "D411", "D213", "D413", "D406", "D407"]
add_ignore = ["D102", "D103"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]


[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
pandas = "^1.4.2"
rich = "^12.5.1"
jupyter = "^1.0.0"
ipywidgets = "^8.0.4"
scikit-learn = "^1.2.1"
gitpython = "^3.1.30"
Cython = "*"
Pillow = "*"
six = "*"
yacs = "*"
torch = "*"
torchvision = "*"
pip = "*"
vidgear = { version = "0.3.0", extras = ["core"] }
jsonargparse = { version = "4.20.0", extras = ["signatures"] }
ultralytics = "*"
seedir = "*"
py-cpuinfo = "*"
black = "*"
ffmpeg-python = "^0.2.0"
pytorch_lightning = "*"
einops = "^0.6.0"
protobuf = "^3.20.0"
ipython = "^8.5.0"
labelbox = "^3.35.0"
filterpy = "^1.4.5"
python-dotenv = "^0.21.1"
opencv-contrib-python = "*"
numpy = "^1.22.4"
scikit-video = "^1.1.11"
poetry = "^1.3.1"
tensorboard = "^2.11.2"
docformatter = "^1.5.1"
scipy = "^1.8.1"
kaggle = "^1.5.12"
mypy = "^0.960"
optuna = "^3.1.1"
networkx = "^3.1"
fire = "^0.5.0"
ndjson = "^0.3.1"
mplsoccer = "^1.1.10"
prospector = "^1.7.7"
loguru = "^0.6.0"
pyroma = "^4.0"
omegaconf = "^2.2.2"
types-python-dateutil = "^2.8.17"
motpy = "^0.0.10"
clip = {git = "https://github.com/openai/CLIP.git"}
gdown = "^4.7.1"

[tool.poetry.dev-dependencies]
papermill = "^2.4.0"
pytest-xdist = "^3.3.1"

[tool.semantic_release]
version_variable = "pyproject.toml:version"


